fix issue #133

+++ src/Form/Field/File.php
@@ -13,23 +13,59 @@ class File extends Field
     const ACTION_KEEP = 0;
     const ACTION_REMOVE = 1;
 
+    /**
+     * Upload directory.
+     *
+     * @var string
+     */
     protected $directory = '';
 
+    /**
+     * File name.
+     *
+     * @var null
+     */
     protected $name = null;
 
+    /**
+     * Options for file-upload plugin.
+     *
+     * @var array
+     */
     protected $options = [];
 
+    /**
+     * Storage instance.
+     *
+     * @var string
+     */
     protected $storage = '';
 
+    /**
+     * Css.
+     *
+     * @var array
+     */
     protected static $css = [
         '/packages/admin/bootstrap-fileinput/css/fileinput.min.css',
     ];
 
+    /**
+     * Js.
+     *
+     * @var array
+     */
     protected static $js = [
         '/packages/admin/bootstrap-fileinput/js/plugins/canvas-to-blob.min.js',
         '/packages/admin/bootstrap-fileinput/js/fileinput.min.js',
     ];
 
+    /**
+     * Create a new File instance.
+     *
+     * @param string $column
+     * @param array  $arguments
+     */
     public function __construct($column, $arguments = [])
     {
         $this->initOptions();
@@ -38,6 +74,21 @@ class File extends Field
         parent::__construct($column, $arguments);
     }
 
+    /**
+     * Initialize the storage instance.
+     *
+     * @return void.
+     */
+    protected function initStorage()
+    {
+        $this->storage = Storage::disk(config('admin.upload.disk'));
+    }
+
+    /**
+     * Initialize file-upload plugin.
+     *
+     * @return void.
+     */
     protected function initOptions()
     {
         $this->options = [
@@ -47,16 +98,36 @@ class File extends Field
         ];
     }
 
-    protected function initStorage()
+    /**
+     * Set options for file-upload plugin.
+     *
+     * @param array $options
+     * @return $this
+     */
+    public function options($options = [])
     {
-        $this->storage = Storage::disk(config('admin.upload.disk'));
+        $this->options = array_merge($this->options, $options);
+
+        return $this;
     }
 
+    /**
+     * Default store path for file upload.
+     *
+     * @return mixed
+     */
     public function defaultStorePath()
     {
         return config('admin.upload.directory.file');
     }
 
+    /**
+     * Specify the directory and name for uplaod file.
+     *
+     * @param string      $directory
+     * @param null|string $name
+     * @return $this
+     */
     public function move($directory, $name = null)
     {
         $this->directory = $directory;
@@ -66,6 +137,12 @@ class File extends Field
         return $this;
     }
 
+    /**
+     * Prepare for saving.
+     *
+     * @param UploadedFile $file
+     * @return mixed|string
+     */
     public function prepare(UploadedFile $file = null)
     {
         if (is_null($file)) {
@@ -84,7 +161,9 @@ class File extends Field
     }
 
     /**
-     * @param $file
+     * Upload file and delete original file.
+     *
+     * @param UploadedFile $file
      *
      * @return mixed
      */
@@ -101,6 +180,11 @@ class File extends Field
         return $target;
     }
 
+    /**
+     * Preview html for file-upload plugin.
+     *
+     * @return string
+     */
     protected function preview()
     {
         $fileName = basename($this->value);
@@ -117,13 +201,12 @@ class File extends Field
 EOT;
     }
 
-    public function options($options = [])
-    {
-        $this->options = array_merge($this->options, $options);
-
-        return $this;
-    }
-
+    /**
+     * Get file visit url.
+     *
+     * @param $path
+     * @return string
+     */
     public function objectUrl($path)
     {
         if (Str::startsWith($path, ['http://', 'https://'])) {
@@ -133,6 +216,11 @@ EOT;
         return trim(config('admin.upload.host'), '/').'/'.trim($path, '/');
     }
 
+    /**
+     * Render file upload field.
+     *
+     * @return \Illuminate\Contracts\View\Factory|\Illuminate\View\View
+     */
     public function render()
     {
         $this->options['initialCaption'] = basename($this->value);
@@ -175,6 +263,8 @@ EOT;
     }
 
     /**
+     * If name already exists, rename it.
+     *
      * @param $file
      *
      * @return void
@@ -186,8 +276,15 @@ EOT;
         }
     }
 
+    /**
+     * Destroy original file.
+     *
+     * @return void.
+     */
     public function destroy()
     {
-        $this->storage->delete($this->original);
+        if ($this->storage->exists($this->original)) {
+            $this->storage->delete($this->original);
+        }
     }
 }
